+++ curiosity/src/main/java/com/hello/curiosity/compose/ui/components/buttons/Button.kt
@@ -8,24 +8,19 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonColors
-import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.ContentAlpha
 import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.material.contentColorFor
-import androidx.compose.material.primarySurface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.compositeOver
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.hello.curiosity.compose.ui.components.buttons.ButtonDefaults.buttonColors
 import com.hello.curiosity.compose.ui.theme.ThemeImpl
 
 typealias action = () -> Unit
@@ -34,21 +29,6 @@ private val btnStyle = ThemeImpl.typography.h5.copy(fontSize = 18.sp)
 private val btnShape = CircleShape
 private val horizontalPadding = 25.dp
 private val verticalPadding = 16.dp
-
-@Composable
-fun buttonColors(
-    backgroundColor: Color = MaterialTheme.colors.primarySurface,
-    contentColor: Color = contentColorFor(backgroundColor),
-    disabledBackgroundColor: Color = MaterialTheme.colors.onSurface.copy(alpha = 0.12f)
-        .compositeOver(MaterialTheme.colors.surface),
-    disabledContentColor: Color = MaterialTheme.colors.onSurface
-        .copy(alpha = ContentAlpha.disabled)
-): ButtonColors = ButtonDefaults.buttonColors(
-    backgroundColor = backgroundColor,
-    contentColor = contentColor,
-    disabledBackgroundColor = disabledBackgroundColor,
-    disabledContentColor = disabledContentColor
-)
 
 @Composable
 fun TextButton(
@@ -118,13 +98,6 @@ fun TextIconButton(
             style = style,
         )
     }
-}
-
-@Composable
-private fun ButtonColors.tintColor(enabled: Boolean) = if (enabled) {
-    contentColor(enabled = true).value
-} else {
-    MaterialTheme.colors.onSurface.copy(alpha = ContentAlpha.disabled)
 }
 
 @Composable
