+++ curiosity/src/main/java/com/hello/curiosity/compose/ui/components/buttons/ButtonDefaults.kt
@@ -1,0 +1,35 @@
+package com.hello.curiosity.compose.ui.components.buttons
+
+import androidx.compose.material.ButtonColors
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.ContentAlpha
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.contentColorFor
+import androidx.compose.material.primarySurface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.compositeOver
+
+object ButtonDefaults {
+    @Composable
+    fun buttonColors(
+        backgroundColor: Color = MaterialTheme.colors.primarySurface,
+        contentColor: Color = contentColorFor(backgroundColor),
+        disabledBackgroundColor: Color = MaterialTheme.colors.onSurface.copy(alpha = 0.12f)
+            .compositeOver(MaterialTheme.colors.surface),
+        disabledContentColor: Color = MaterialTheme.colors.onSurface
+            .copy(alpha = ContentAlpha.disabled)
+    ): ButtonColors = ButtonDefaults.buttonColors(
+        backgroundColor = backgroundColor,
+        contentColor = contentColor,
+        disabledBackgroundColor = disabledBackgroundColor,
+        disabledContentColor = disabledContentColor
+    )
+}
+
+@Composable
+internal fun ButtonColors.tintColor(enabled: Boolean) = if (enabled) {
+    contentColor(enabled = true).value
+} else {
+    MaterialTheme.colors.onSurface.copy(alpha = ContentAlpha.disabled)
+}
